requirements.txt
anthropic>=0.21.0
tweepy>=4.14.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
feedparser>=6.0.0
asyncio
pathlib
json5
